# core/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/core/api/v1/users/{userId}/bills/{billId}/transactions": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "summary": "Посмотреть историю транзакций по счету у пользователя",
        "description": "Описание прав:\n* TRANSACTION_READ - доступ к просмотру истории транзакций своего счета\n* TRANSACTION_READ_OTHERS - доступ к просмотру истории транзакций любого счета\n",
        "operationId": "getTransactionsV1",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "billId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "User Authentication": []
          }
        ]
      },
      "post": {
        "tags": [
          "transaction-controller"
        ],
        "summary": "Выполнить транзакцию по счету у пользователя",
        "description": "Описание прав:\n* TRANSACTION_WRITE - доступ к выполнению транзакции на своем счете\n* TRANSACTION_WRITE_OTHERS - доступ к выполнению транзакции на любом счете\n",
        "operationId": "createTransactionV1",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "billId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "User Authentication": []
          }
        ]
      }
    },
    "/users/{userId}/bills/{billId}/transactions": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "summary": "Посмотреть историю транзакций по счету у пользователя",
        "description": "Описание прав:\n* TRANSACTION_READ - доступ к просмотру истории транзакций своего счета\n* TRANSACTION_READ_OTHERS - доступ к просмотру истории транзакций любого счета\n",
        "operationId": "getTransactionsV1_1",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "billId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "User Authentication": []
          }
        ]
      },
      "post": {
        "tags": [
          "transaction-controller"
        ],
        "summary": "Выполнить транзакцию по счету у пользователя",
        "description": "Описание прав:\n* TRANSACTION_WRITE - доступ к выполнению транзакции на своем счете\n* TRANSACTION_WRITE_OTHERS - доступ к выполнению транзакции на любом счете\n",
        "operationId": "createTransactionV1_1",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "billId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "User Authentication": []
          }
        ]
      }
    },
...
    "/internal/bills/{billId}/transactions": {
      "post": {
        "tags": [
          "internal"
        ],
        "summary": "Выполнить транзакцию по счету у пользователя",
        "description": "Этот метод предназначен для внутреннего сервиса \u0027Кредиты\u0027",
        "operationId": "createTransaction",
        "parameters": [
          {
            "name": "billId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "Internal Authentication": []
          }
        ]
      }
    },
    "/core/api/v1/internal/bills/{billId}/transactions": {
      "post": {
        "tags": [
          "internal"
        ],
        "summary": "Выполнить транзакцию по счету у пользователя",
        "description": "Этот метод предназначен для внутреннего сервиса \u0027Кредиты\u0027",
        "operationId": "createTransaction_1",
        "parameters": [
          {
            "name": "billId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "Internal Authentication": []
          }
        ]
      }
    },
...
    "/core/api/v2/users/{userId}/bills/{billId}/transactions": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "summary": "Посмотреть историю транзакций по счету у пользователя",
        "operationId": "getUserTransactionsV2",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "billId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionDto"
                  }
                }
              }
            }
          }
        },
        "deprecated": true
      }
    },
...
    "/core/api/v2/bills/{billId}/transactions": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "summary": "Посмотреть историю транзакций по счету у пользователя",
        "operationId": "getTransactionsV2",
        "parameters": [
          {
            "name": "billId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionDto"
                  }
                }
              }
            }
          }
        },
        "deprecated": true
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "Код ошибки. Рекомендуется использовать тип ошибки вместо кода.",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "description": "Тип ошибки",
            "enum": [
              "UNKNOWN",
              "BAD_GATEWAY",
              "CANNOT_NEGATIVE_BILL_BALANCE",
              "CLOSING_BILL_WITH_POSITIVE_BALANCE",
              "CLOSING_BILL_WITH_NEGATIVE_BALANCE",
              "TRANSACTION_WITH_ZERO_BALANCE_CHANGE",
              "CLOSING_SYSTEM_BILL",
              "BILL_NOT_FOUND",
              "CLOSED_BILL_OPERATION"
            ]
          },
          "message": {
            "type": "string",
            "description": "Сообщение об ошибке",
            "example": "Сообщение об ошибке"
          }
        }
      },
...
      "TransactionCreationDto": {
        "type": "object",
        "properties": {
          "balanceChange": {
            "type": "integer",
            "description": "Изменение баланса счета в копейках",
            "format": "int64"
          }
        }
      },
      "TransactionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор транзакции",
            "format": "uuid"
          },
          "billId": {
            "type": "string",
            "description": "Идентификатор счета транзакции",
            "format": "uuid"
          },
          "balanceChange": {
            "type": "integer",
            "description": "Изменение баланса счета в копейках",
            "format": "int64"
          },
          "reason": {
            "type": "string",
            "description": "Причина выполнения транзакции",
            "enum": [
              "TERMINAL",
              "LOAN",
              "TRANSFER"
            ]
          },
          "instant": {
            "type": "string",
            "description": "Момент времени выполнения транзакции",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}